+++ saleor/warehouse/models.py
@@ -25,7 +25,9 @@ class Warehouse(models.Model):
     slug = models.SlugField(max_length=255, unique=True)
     company_name = models.CharField(blank=True, max_length=255)
 
-    shipping_zones = models.ManyToManyField(ShippingZone, blank=True)
+    shipping_zones = models.ManyToManyField(
+        ShippingZone, blank=True, related_name="warehouses"
+    )
     address = models.ForeignKey(Address, on_delete=models.PROTECT)
 
     email = models.EmailField(blank=True, default="")